rm frontend calc fns; organize remaining utils
